+++ llvm/lib/Target/AMDGPU/MCTargetDesc/AMDGPUInstPrinter.cpp
@@ -183,6 +183,11 @@ void AMDGPUInstPrinter::printCPol(const MCInst *MI, unsigned OpNo,
     if (Imm & CPol::NV)
       O << " nv";
 
+    if (AMDGPU::isGFX13(STI)) {
+      const int64_t CFS = Imm & CPol::CFS;
+      printCFS(MI, CFS, O);
+    }
+
     return;
   }
 
@@ -275,6 +280,30 @@ void AMDGPUInstPrinter::printScope(int64_t Scope, raw_ostream &O) {
     O << "SCOPE_SYS";
   else
     llvm_unreachable("unexpected scope policy value");
+}
+
+void AMDGPUInstPrinter::printCFS(const MCInst *MI, int64_t CFS,
+                                 raw_ostream &O) {
+  if (CFS == CPol::CFS_256B)
+    return;
+
+  O << " cfs:";
+
+  switch (CFS) {
+  case CPol::CFS_128B:
+    O << "CFS_128B";
+    break;
+  case CPol::CFS_64B:
+    O << "CFS_64B";
+    break;
+  case CPol::CFS_32B:
+    O << "CFS_32B";
+    break;
+  default:
+    llvm_unreachable("unexpected cache fill size value");
+  }
+
+  return;
 }
 
 void AMDGPUInstPrinter::printDim(const MCInst *MI, unsigned OpNo,